feat: implement synchronize block process
