+++ chain/context/chain_context.go
@@ -32,6 +32,7 @@ const (
 	PovService         = "povService"
 	MinerService       = "minerService"
 	AutoReceiveService = "autoReceiveService"
+	RollbackService    = "rollbackService"
 	MetricsService     = "metricsService"
 )
