+++ chain/helper.go
@@ -34,6 +34,9 @@ func RegisterServices(cc *context.ChainContext) error {
 		_ = cc.Register(context.WalletService, walletService)
 	}
 
+	rollbackService := NewRollbackService(cfgFile)
+	_ = cc.Register(context.RollbackService, rollbackService)
+
 	if len(cfg.P2P.BootNodes) > 0 {
 		netService, err := NewP2PService(cfgFile)
 		if err != nil {
